Unificados los metodos en un unico fichero
